Remove old MultiButton. Remove scss and reach from button
