+++ .storybook/designmanual.scss
@@ -4,7 +4,6 @@
 @import '~@ndla/tabs/scss/tabs';
 @import '~@ndla/core/scss/utilities';
 @import '~@ndla/video-search/scss/video-search';
-@import '~@ndla/button/src/all.scss';
 
 .c-componentinfo {
   > .c-tabs {
