+++ packages/button/package.json
@@ -7,8 +7,7 @@
   "module": "es/index.js",
   "types": "lib/index.d.ts",
   "scripts": {
-    "build:css": "webpack --config ../../css.webpack.config.js && cp ./dist/*.css ./es && cp ./dist/*.css ./lib",
-    "build": "concurrently \"yarn build:types\" \"yarn build:css\"",
+    "build": "yarn build:types",
     "build:types": "tsc -p tsconfig.build.json",
     "prepublish": "yarn build"
   },
@@ -23,18 +22,14 @@
   ],
   "author": "ndla@knowit.no",
   "files": [
-    "scss",
     "lib",
-    "es",
-    "dist"
+    "es"
   ],
   "dependencies": {
     "@ndla/core": "^3.0.3",
     "@radix-ui/react-dropdown-menu": "^2.0.1",
     "@ndla/icons": "^2.1.0",
     "@ndla/tooltip": "^3.1.1",
-    "@reach/menu-button": "^0.16.2",
-    "@reach/popover": "^0.16.2",
     "focus-trap-react": "^8.9.2"
   },
   "peerDependencies": {
