+++ packages/button/src/MultiButton.stories.tsx
@@ -10,7 +10,7 @@ import styled from '@emotion/styled';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 import React from 'react';
 import { defaultParameters } from '../../../stories/defaults';
-import MultiButton from './MultiButtonV2';
+import MultiButton from './MultiButton';
 
 const Wrapper = styled.div`
   margin: 5em;
