--- packages/button/src/all.scss
@@ -1,1 +1,0 @@
-@import '~@reach/menu-button/styles.css';
