+++ packages/ndla-ui/src/all.scss
@@ -4,4 +4,3 @@
 @import '~@ndla/tabs/scss/tabs';
 @import '~@ndla/core/scss/utilities';
 @import '~@ndla/video-search/scss/video-search';
-@import '~@ndla/button/src/all.scss';
